+++ test/spec/feature_layer.js
@@ -1,4 +1,4 @@
-describe('L.mapbox.markerLayer', function() {
+describe('L.mapbox.featureLayer', function() {
     'use strict';
 
     var server;
@@ -12,7 +12,7 @@ describe('L.mapbox.markerLayer', function() {
     });
 
     it('loads data from a GeoJSON source', function() {
-        var layer = L.mapbox.markerLayer(helpers.geoJson),
+        var layer = L.mapbox.featureLayer(helpers.geoJson),
             marker = layer.getLayers()[0];
         expect(marker instanceof L.Marker).to.equal(true);
         expect(marker.getLatLng()).to.be.near({lng: -77.0203, lat: 38.8995}, 0);
@@ -20,7 +20,7 @@ describe('L.mapbox.markerLayer', function() {
 
     it('loads data from a GeoJSON URL', function() {
         var url = 'http://api.tiles.mapbox.com/v3/examples.map-zr0njcqy/markers.geojson',
-            layer = L.mapbox.markerLayer(url);
+            layer = L.mapbox.featureLayer(url);
 
         server.respondWith("GET", url,
             [200, { "Content-Type": "application/json" }, JSON.stringify(helpers.geoJson)]);
@@ -34,7 +34,7 @@ describe('L.mapbox.markerLayer', function() {
     it('loads data for a map ID', function() {
         var id = 'examples.map-zr0njcqy',
             url = 'http://a.tiles.mapbox.com/v3/examples.map-zr0njcqy/markers.geojson',
-            layer = L.mapbox.markerLayer(id);
+            layer = L.mapbox.featureLayer(id);
 
         server.respondWith("GET", url,
             [200, { "Content-Type": "application/json" }, JSON.stringify(helpers.geoJson)]);
@@ -47,7 +47,7 @@ describe('L.mapbox.markerLayer', function() {
 
     it('replaces jsonp URLs with the equivalent json URL', function() {
         var url = 'http://api.tiles.mapbox.com/v3/examples.map-zr0njcqy/markers.geojson',
-            layer = L.mapbox.markerLayer(url + 'p');
+            layer = L.mapbox.featureLayer(url + 'p');
 
         server.respondWith("GET", url,
             [200, { "Content-Type": "application/json" }, JSON.stringify(helpers.geoJson)]);
@@ -60,12 +60,12 @@ describe('L.mapbox.markerLayer', function() {
 
     describe("#loadURL", function() {
         it('returns self', function() {
-            var layer = L.mapbox.markerLayer();
+            var layer = L.mapbox.featureLayer();
             expect(layer.loadURL('http://a.tiles.mapbox.com/v3/mapbox.map-0l53fhk2.json')).to.eql(layer);
         });
 
         it('emits a ready event', function(done) {
-            var layer = L.mapbox.markerLayer();
+            var layer = L.mapbox.featureLayer();
 
             layer.on('ready', function() {
                 done();
@@ -79,7 +79,7 @@ describe('L.mapbox.markerLayer', function() {
         });
 
         it('emits an error event', function(done) {
-            var layer = L.mapbox.markerLayer();
+            var layer = L.mapbox.featureLayer();
 
             layer.on('error', function(e) {
                 expect(this).to.equal(layer);
@@ -97,13 +97,20 @@ describe('L.mapbox.markerLayer', function() {
 
     describe("#setGeoJSON", function() {
         it("sets GeoJSON", function() {
-            var layer = L.mapbox.markerLayer();
+            var layer = L.mapbox.featureLayer();
             expect(layer.setGeoJSON(helpers.geoJson)).to.eql(layer);
             expect(layer.getGeoJSON()).to.eql(helpers.geoJson);
         });
-
+        it("styles GeoJSON features", function(done) {
+            var layer = L.mapbox.featureLayer();
+            expect(layer.setGeoJSON(helpers.geoJsonPoly)).to.eql(layer);
+            layer.eachLayer(function(l) {
+                expect(l.options.color).to.eql('#f00');
+                done();
+            });
+        });
         it("removes existing layers", function() {
-            var layer = L.mapbox.markerLayer(helpers.geoJson);
+            var layer = L.mapbox.featureLayer(helpers.geoJson);
             layer.setGeoJSON([]);
             expect(layer.getLayers()).to.be.empty();
         });
@@ -112,14 +119,14 @@ describe('L.mapbox.markerLayer', function() {
     describe("#getFilter", function() {
         it("returns the filter option when not given an argument", function() {
             var filter = function () {},
-                layer = L.mapbox.markerLayer(null, {filter: filter});
+                layer = L.mapbox.featureLayer(null, {filter: filter});
             expect(layer.getFilter()).to.equal(filter);
         });
     });
 
     describe("#setFilter", function() {
         it("filters features to those for which the function returns true", function() {
-            var layer = L.mapbox.markerLayer(helpers.geoJson);
+            var layer = L.mapbox.featureLayer(helpers.geoJson);
 
             var fooFilter = function (f) { return f.properties.title === 'foo'; };
             expect(layer.setFilter(fooFilter)).to.eql(layer);
@@ -144,13 +151,13 @@ describe('L.mapbox.markerLayer', function() {
     };
 
     it('sanitizes marker content', function() {
-        var layer = L.mapbox.markerLayer(unsanitary);
+        var layer = L.mapbox.featureLayer(unsanitary);
 
         expect(layer.getLayers()[0]._popup._content).not.to.match(/<script>/);
     });
 
     it('supports a custom sanitizer', function() {
-        var layer = L.mapbox.markerLayer(unsanitary, {
+        var layer = L.mapbox.featureLayer(unsanitary, {
             sanitizer: function(_) { return _; }
         });